+++ docs/about/release-notes.md
@@ -17,6 +17,7 @@ To fix incoherences or potential limitation in JSON graphs :
 ### Functionalities
 - Downloader: add a new provider from URL (use --url-provider)
 - Add Documentation "how to add a new sensor"
+- Add Landsat AWS Provider (Catalogue&Download)
 
 ### Bug fixes
 - LocalProvider does not require that a date is included in the product name